mysite/requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2.3
django-rest-knox==4.1.0
djangorestframework==3.12.4
psycopg2==2.8.6
pycparser==2.20
pytz==2021.1
sqlparse==0.4.1
